+++ resources/classes/voice_info_data.gd
@@ -1,0 +1,7 @@
+extends Resource
+
+class_name VoiceInfoData
+
+@export var voice: AudioStream
+@export var offset_start: float = 0
+@export var offset_end: float = 0