+++ resources/components/clock_voice_playback.gd
@@ -2,54 +2,82 @@ extends Node
 
 class_name ClockVoicePlayback
 
-@export var voices: Dictionary = {}
-@export var voice_word_offset: Dictionary = {}
-@export var slient_time: float = 0.4
+signal finished()
 
-@onready var audio_player_1 = $AudioStreamPlayer
-@onready var audio_player_2 = $AudioStreamPlayer2
-@onready var silent_timer = $SilentTimer
+@export_range(0, 100, 1.0, "or_greater") var volume_percent: float = 150
 
-var voice_queue: Array[String] = []
-var use_alt: bool = false
+@export_group("Playback")
+@export var audio_players: Array[AudioStreamPlayer]
+@export var silent_timer: Timer
+var channel: int = 0
+var token_queue: Array[ReadTextData] = []
+var _playing: bool = false
 
 func _process(delta: float) -> void:
-	if voice_queue.size() > 0 and silent_timer.is_stopped():
-		var token = voice_queue.pop_front()
-		if token == " ": # silent voice
-			silent_timer.start(slient_time)
-		else:
-			var current_voice : AudioStream = voices.get(token)
-			if current_voice:
-				var offset : float = 0
-				if voice_queue.size() > 0 and voice_queue[0] != " ":
-					offset = voice_word_offset.get(token, offset)
-				var time_left = max(current_voice.get_length() + offset, 0)
-				print("voice_info => ", {
-					"token": token,
-					"voice_len": current_voice.get_length(),
-					"offset": offset,
-					"time_left": time_left,
-				})
-				
-				var audio_player = _get_audio_player()
-				audio_player.stream = current_voice
-				audio_player.play()
-				use_alt = not use_alt
-				silent_timer.start(time_left)
+	for player in audio_players:
+		player.volume_linear = volume_percent / 100.0
+	
+	# check finished
+	if token_queue.size() == 0 and silent_timer.is_stopped() and _playing:
+		_playing = false
+		finished.emit()
+	
+	if token_queue.size() > 0 and silent_timer.is_stopped():
+		_playing = true
+		_queue_next_voice()
+		
+func _queue_next_voice():
+	var token: ReadTextData = token_queue.pop_front()
+	
+	if token.silent: # silent voice
+		silent_timer.start(token.silent_duration)
+	else:
+		var voice_text = token.get_voice_text()
+		var current_voice : AudioStream
+		var offset_start: float = 0
+		var offset_end: float = 0
+		
+		var resource_path = "res://resources/resources/voice_info_data/%s.tres" % [voice_text]
+		var resource_voice = ResourceLoader.load(resource_path) as VoiceInfoData
+		if resource_voice:
+			current_voice = resource_voice.voice
+			offset_start = resource_voice.offset_start
+			offset_end = resource_voice.offset_end
+		
+		#print_debug("voice_info => ", {
+			#"voice_text": voice_text,
+			#"resource_path": resource_path,
+			#"has_voice": not not resource_voice,
+			#"offset_start": offset_start,
+			#"offset_end": offset_end,
+		#})
+		
+		if current_voice:
+			var voice_duration = max(current_voice.get_length() + offset_end, 0)
+			
+			#print_debug("voice_info => ", {
+				##"voice_len": current_voice.get_length(),
+				##"voice_duration": voice_duration,
+				#"voice_text": voice_text,
+			#})
+			
+			var audio_player = _get_audio_player()
+			audio_player.stream = current_voice
+			audio_player.play(offset_start)
+			channel = (channel + 1) % audio_players.size()
+			silent_timer.start(voice_duration)
 
 func _get_audio_player() -> AudioStreamPlayer:
-	if use_alt:
-		return audio_player_2
-	return audio_player_1
+	return audio_players[channel]
 
-func play(tokens: Array[String]):
+func play(tokens: Array[ReadTextData]):
 	stop()
-	voice_queue = tokens.slice(0)
+	token_queue = tokens.slice(0)
 
 func stop():
-	voice_queue = []
-	audio_player_1.stop()
-	audio_player_2.stop()
+	token_queue = []
 	silent_timer.stop()
+	channel = 0
+	for player in audio_players:
+		player.stop()